First commit
A demo of how to use C const array/structs to derive a pseudo const
NSDictionary, accessed using an enum type. The advantage? The
definition of the values is decoupled from the accessor, so it's
extendable, without editing runtime code.

You can use the technique (minus the structs) to map enum types onto an
array of strings.