Updated tests and docu appropriately, according to the use of FRUIT_TYPES_COUNT
Introducing the use of FRUIT_TYPES_COUNT to define the array size had
many knock on effects, escpeccially in respect to use of a sentinel,
and the "Unknown" enumeration, defined for that purpose. So some minor
code changes, but above all much documentation to explain what's going
on and tests altered to utilise new opportunities to validate.